Split fx quantization related part into fx_int8.py
