+++ torchbenchmark/util/model.py
@@ -14,6 +14,7 @@ from torchbenchmark.util.extra_args import check_correctness_p, parse_opt_args, 
                                            parse_decoration_args, apply_decoration_args, is_staged_train_test, \
                                            TEST_STAGE
 from torchbenchmark.util.env_check import set_random_seed, correctness_check, stableness_check, is_hf_model
+from torchbenchmark.util.fx_int8 import get_sub_module, prepare_sub_module, convert_sub_module
 
 class PostInitProcessor(type):
     def __call__(cls, *args, **kwargs):
@@ -369,105 +370,23 @@ class BenchmarkModel(metaclass=PostInitProcessor):
         if hasattr(self, 'example_inputs'):
             self.example_inputs = inputs_convert(self.example_inputs)
         else:
-            warnings.warn(UserWarning(f"{model_name} example inputs doesn't convert to `channels_last`!"))     
+            warnings.warn(UserWarning(f"{model_name} example inputs doesn't convert to `channels_last`!"))
 
     def enable_fx_int8(self, quant_engine:str='x86'):
-        from torch.ao.quantization import QuantWrapper, get_default_qconfig_mapping, get_default_qconfig_propagation_list
-        from torch.ao.quantization.quantize_fx import _fuse_fx, prepare_fx, convert_fx
         torch.backends.quantized.engine = quant_engine
-        qconfig_mapping = get_default_qconfig_mapping(quant_engine)
-
-        def _append_attr(fx_module, module, fx_white_list=[]):
-            import re
-            fx_attr = dir(fx_module)
-            org_attr = dir(module)
-            ignore_match_patterns = [r"_", r"quant", r"dequant", r"weight",
-                                    r"bias", r'activation_post_process']
-            ignore_search_patterns = [r"_scale_", r"_zero_point_",
-                                    r'_activation_post_process_']
-            add_special_patterns = [r"_forward_hooks", r"_forward_pre_hooks", r"_backward_hooks"]
-            attr_names = []
-            for i in org_attr:
-                if type(module) in fx_white_list and type(module) != torch.nn.Sequential \
-                and any([re.search(p, i) for p in add_special_patterns]):
-                    continue
-                if any([re.search(p, i) for p in add_special_patterns]) \
-                or (i not in fx_attr \
-                    and not any([re.match(p, i) for p in ignore_match_patterns]) \
-                    and not any([re.search(p, i) for p in ignore_search_patterns])) :
-                    attr_names.append(i)
-            for name in attr_names:
-                attr = getattr(module, name, None)
-                if isinstance(attr, torch.nn.Module) or \
-                isinstance(attr, torch.quantization.qconfig.QConfig):
-                    continue
-                setattr(fx_module, name, attr)
-            return fx_module
-
-        def _get_sub_module(model, module_dict, prefix):
-            import transformers
-            fx_white_list = get_default_qconfig_propagation_list()
-            ignore_list = []
-            if is_hf_model:
-                import transformers
-                ignore_list.extend([transformers.models.gpt2.modeling_gpt2.GPT2Attention, transformers.models.t5.modeling_t5.T5DenseActDense])
-            for name, module in model.named_children():
-                quant_wrap_flag = False
-                if type(module) in ignore_list:
-                    continue
-                op_name = prefix + "." + name if prefix != "" else name
-                if op_name not in module_dict:
-                    continue
-                if type(module) in fx_white_list and type(module) != torch.nn.Sequential:
-                    module = QuantWrapper(module)
-                    quant_wrap_flag = True
-                try:
-                    graph_module = torch.fx.symbolic_trace(module)
-                    if not quant_wrap_flag and str(module.get_submodule).count("\n") != str(graph_module.get_submodule).count("\n"):
-                        continue
-                    _fuse_fx(graph_module, False)
-                    setattr(model, name, module)
-                    sub_module_list.append(op_name)
-                except:
-                    _get_sub_module(module, module_dict, op_name)
-            return model
-
-        def _prepare_sub_module(sub_module_list, model, prefix):
-            for name, module in model.named_children():
-                op_name = prefix + '.' + name if prefix != '' else name
-                if op_name in sub_module_list:
-                    prepared_module = prepare_fx(module, qconfig_mapping, None)
-                    _append_attr(prepared_module, module)
-                    setattr(model, name, prepared_module)
-                else:
-                    _prepare_sub_module(sub_module_list, module, op_name)
-            return model
-
-        def _convert_sub_module(sub_module_list, model, prefix):
-            for name, module in model.named_children():
-                op_name = prefix + '.' + name if prefix != '' else name
-                if op_name in sub_module_list:
-                    convert_module = convert_fx(module)
-                    setattr(model, name, convert_module)
-                else:
-                    _convert_sub_module(sub_module_list, module, op_name)
-            return model
-
         try:
             model, _ = self.get_module()
-            sub_module_list = []
             # Get sub modules
-            model = _get_sub_module(model, dict(model.named_modules()), '')
+            model, sub_module_list = get_sub_module(model, dict(model.named_modules()), '')
             if not len(sub_module_list):
                 warnings.warn(UserWarning(f"{self.name} doesn't have submodule can ben quantized!"))
-            model = _prepare_sub_module(sub_module_list, model, '')
+            model = prepare_sub_module(sub_module_list, model, '', quant_engine)
             self.set_module(model)
             # Calibration
             self.eval()
             model, _ = self.get_module()
-            model = _convert_sub_module(sub_module_list, model, '')
+            model = convert_sub_module(sub_module_list, model, '')
             self.set_module(model)
         except Exception as e:
             print(e)
-            warnings.warn(UserWarning(f"{self.name} doesn't support `fx_int8` yet!"))
-            raise RuntimeError
+            raise RuntimeError(f"{self.name} doesn't support `fx_int8` yet!")
